Inserting pipex in exec
Normed minishell.h and added main_bonus from pipex

Libft makefile refactor, AGAIN. Some norm improvements made, working on command_creator trying to resolve a bug with quoting

Fixed -e error with echo in Makefile && created pipex modification directory

Destroying pipex main

"plugged" pipex, no path?

mytmux

transfer done

Trying to get child exit code in parent

transfer of script

Transfer done

IT WORKS

- Add of ft_strlcat in libft.h (forgotten for too long, mb)
- Command creation repaired
- Execute simple command succesfully

temp

Makefile patch, was ugly ah

Command not found exit code (127)

Big changes:
 - Parser rebuild
 - Builtin functions refactor for t_command struct compatibility
 - File redirection implemented (unstable ?)
 - Builtins plugged in execution (only as sole command)
 - 42 norm conform, only main is ugly

Norm (removed ternaries)

env_var_delete -> env_var_free

ft_memdup in command_creator

Infinite loop in command_creator solved (was token_dup->next not NULL)

Removed debug messages

Execution pipeline created, waiting for parsing compatibility

Solved pipe lexing problem and check for consecutive pipe error

Echo builtin refactor to work with t_command struct
Working on lexer problems (syntax errors)

< file cat | ls > file gives fd error?

Add of ft_strdup of token->str in arg_redirect duplication

+++ libft/include/vector.h
@@ -6,7 +6,7 @@
 /*   By: ll-hotel <ll-hotel@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/01 15:24:46 by ll-hotel          #+#    #+#             */
-/*   Updated: 2024/04/16 16:36:18 by ll-hotel         ###   ########.fr       */
+/*   Updated: 2024/05/07 10:42:32 by ll-hotel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@ typedef struct s_vector	t_vec;
 
 struct	s_vector
 {
-	void			*array;
-	unsigned long	size;
-	unsigned long	elem_size;
-	unsigned long	allocated_size;
+	void				*array;
+	unsigned long long	size;
+	unsigned long long	elem_size;
+	unsigned long long	allocated_size;
 };
 
 void	vec_new(t_vec *vec, unsigned long elem_size);
